app/requirements.txt
fastapi
uvicorn
langchain
langchain-core
langchain-openai
langchain-community
python-dotenv
pymongo
pinecone
langchain-pinecone
selenium
webdriver-manager
upstash-redis
